pyproject.toml
[project]
name = "ai4drugdesign-spring2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.4.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]
